Add hooks to VSPSelect
