+++ app/components/inputs/VSPSelect/VSPSelect.modules.css
@@ -1,0 +1,3 @@
+.error {
+    color: var(--error-text-color);
+}
